#2231 move builders to corresponding packages
