+++ vuu-java/src/test/java/org/finos/vuu/api/JoinTableDefBuilderTest.java
@@ -1,12 +1,5 @@
-package org.finos.vuu.util;
+package org.finos.vuu.api;
 
-import org.finos.vuu.api.JoinSpec;
-import org.finos.vuu.api.JoinTableDef;
-import org.finos.vuu.api.JoinTo;
-import org.finos.vuu.api.LeftOuterJoin$;
-import org.finos.vuu.api.Link;
-import org.finos.vuu.api.TableDef;
-import org.finos.vuu.api.TableVisibility;
 import org.finos.vuu.core.table.Column;
 import org.finos.vuu.core.table.RangeSettings;
 import org.finos.vuu.core.table.SimpleColumn;
